Revert "Use correct integer type also on 32-bit systems1"
This reverts commit 17d244e0d6073c1fef63bc2b189735293df265bd.

+++ src/ricci.jl
@@ -15,7 +15,7 @@ export to_std_string
 struct Upper end
 struct Lower end
 
-Letter = Int
+Letter = Int64
 
 LowerOrUpperIndex = Union{Lower, Upper}
 
@@ -113,7 +113,7 @@ function ==(left::Zero, right::Zero)
     return left.indices == right.indices
 end
 
-ContractingPair = Tuple{Int, Int}
+ContractingPair = Tuple{Int64, Int64}
 Contractions = Vector{ContractingPair}
 
 # TODO: Rename to contraction
